+++ CHANGELOG.md
@@ -2561,3 +2561,4 @@
 - [2026-07-03 21:39] test(auth): implement widget integration tests for auth
 - [2026-07-03 22:07] feat(storage): add bi-directional event bus handler for storage
 - [2026-07-03 22:36] style(test): harmonize light and dark theme palette for test
+- [2026-07-03 22:57] style(forms): harmonize light and dark theme palette for forms